Update likePainting.html
Ficed:
The error occurs because AWS.config.credentials.identityId is being accessed in the loadLikedPaintings function when the user is not authenticated, resulting in a null value for AWS.config.credentials. To fix this, you need to check if the credentials are loaded and valid before attempting to access identityId.